Move back function InitializeSecurityPolicy to firewall.go
Move back function InitializeSecurityPolicy to firewall.go as branch main does.
